# gh-124688: _decimal: Get a module state from ctx objects for performance