<!-- gh-issue-number: gh-124688 -->
* Issue: gh-124688
<!-- /gh-issue-number -->