@vstinner 2024-09-27
Do you have a benchmark showing the effect of this change?