@neonene 2024-09-27
`_dec_mpd_radix()` is added (818c24fd2c96d8e1494528397a72c14352be5bd6) so as not to be passed a different object, not for performance.