@neonene 2024-09-27
On release builds, the result was as fast as the global state access: https://github.com/python/cpython/pull/123100#issuecomment-2298566533 (`PR with alternatives`: 1.08x  )