@neonene 2024-09-27
> Would you mind to share your results (before/after, difference)?
Wait please.