@neonene 2024-09-27
Picked a few `telco` results on Windows PGO builds:

a) main with d69529d31ccd1510843cfac1ab53bb8cb027541f restored

```
9.68 ms +- 0.17 ms -> 8.94 ms +- 0.15 ms: 1.08x faster  # PR (ctx object)
9.68 ms +- 0.17 ms -> 8.76 ms +- 0.11 ms: 1.10x faster  # global state access
```

b) (a) with some patches

```
8.99 ms +- 0.11 ms -> 8.49 ms +- 0.16 ms: 1.06x faster  # PR (ctx object)
8.99 ms +- 0.11 ms -> 8.44 ms +- 0.20 ms: 1.07x faster  # global state access
```

c) (b) with 1 diff in `PyType_GetBaseByToken()`

```
8.91 ms +- 0.21 ms -> 8.78 ms +- 0.16 ms: 1.02x faster  # PR (ctx object)
8.91 ms +- 0.21 ms -> 8.25 ms +- 0.17 ms: 1.08x faster  # global state access
```