@vstinner 2024-09-28
Sorry, I don't know how to read your 6 benchmarks. I don't understand what you compared. I don't know what are "some patches" nor why you mention PyType_GetBaseByToken(), it doesn't help me to compare *this* PR.

So I just compared the main branch to this PR:

```
Mean +- std dev: [ref] 8.18 ms +- 0.28 ms -> [change] 7.53 ms +- 0.18 ms: 1.09x faster
```

Nice! 1.09x faster is worth it.