@vstinner 2024-09-28
Merged, thanks @neonene!